Fixed Issues with sorting and added a MVP UI design for such so. Updated exoplayer to improve Stability  I broke dimmer windows though
